fix(runner): refuse the Windows user data dir instead of writing it unprotected
editor_windows.go kept the path-based MkdirAll and WriteFile the unix side just
replaced, with a header comment saying that path needs a junction-safe
implementation before editor tunnels are enabled here. A comment is not a
guard. Whoever flips editorTunnelSupported gets the unprotected path running,
and the only thing standing in their way is a test about tunnel support that
says nothing about this file.

Both functions now return an error. Nothing reachable changes: editor tunnels
are rejected on Windows in validateClientTypeRequirements, and chownTreeNoFollow
sat behind os.Getuid() == 0, which Windows never satisfies. What changes is what
happens the day someone lifts that rejection — a red test naming this file,
rather than a write through whatever the path resolves to. The agent runs as
LocalSystem there, and a directory junction, which any user can create without
the privilege a symlink requires, redirects a path-based write the way a symlink
does on unix.

TestSetupUserDataDir and TestSetupUserDataDirIdempotent move to
editor_unix_test.go with that. They assert files landing on disk, which is a
unix behavior now, and editor_test.go carries no build tag so they ran on the
Windows job. TestSetupUserDataDirRejectsUnsafeDirName stays: the name check runs
before either platform implementation and holds on both.

Reported by Copilot on #457.

Constraint: editor_test.go has no build tag, so anything in it asserting a written file has to move once one platform stops writing
Rejected: leaving the implementation and adding a Windows test that pins editorTunnelSupported — that test already exists in tunnel_test.go and points at the tunnel handler, not at the file that needs the work
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: implementing the junction-safe Windows path (#379) means deleting errUserDataDirUnsupported and its test, not editing around them
